Extract tests for isDataView to own file
