+++ lib/referee.test.js
@@ -4,45 +4,6 @@ var referee = require("../lib/referee");
 var testHelper = require("./test-helper");
 var sinon = require("sinon");
 
-testHelper.assertionTests("assert", "isDataView", function(
-    pass,
-    fail,
-    msg,
-    error
-) {
-    function captureArgs() {
-        return arguments;
-    }
-
-    var ab = new global.ArrayBuffer(8);
-    var dv = new global.DataView(ab);
-
-    pass("for DataView", dv);
-    fail("for ArrayBuffer", ab);
-    fail("for array", []);
-    fail("for object", {});
-    fail("for arguments", captureArgs());
-    msg(
-        "fail with descriptive message",
-        "[assert.isDataView] Expected {  } to be a DataView",
-        {}
-    );
-    msg(
-        "fail with custom message",
-        "[assert.isDataView] Nope: Expected {  } to be a DataView",
-        {},
-        "Nope"
-    );
-    error(
-        "for object",
-        {
-            code: "ERR_ASSERTION",
-            operator: "assert.isDataView"
-        },
-        {}
-    );
-});
-
 testHelper.assertionTests("assert", "isDate", function(pass, fail, msg, error) {
     function captureArgs() {
         return arguments;
